fix: sos to single button
